refactor: Move enums to domain-specific packages and add UserRepository
