--- apps/backend/src/main/java/com/sos/backend/domain/common/enums/Role.java
@@ -1,5 +1,0 @@
-package com.sos.backend.domain.common.enums;
-
-public enum Role {
-    USER, ADMIN
-}
