+++ apps/backend/src/main/java/com/sos/backend/domain/email/enums/VerificationPurpose.java
@@ -1,4 +1,4 @@
-package com.sos.backend.domain.common.enums;
+package com.sos.backend.domain.email.enums;
 
 public enum VerificationPurpose {
     SIGNUP, RESET_PASSWORD
