+++ apps/backend/src/main/java/com/sos/backend/domain/user/enums/CommunicateChannel.java
@@ -1,4 +1,4 @@
-package com.sos.backend.domain.common.enums;
+package com.sos.backend.domain.user.enums;
 
 import java.util.Arrays;
