+++ apps/backend/src/main/java/com/sos/backend/domain/user/enums/Gender.java
@@ -1,0 +1,5 @@
+package com.sos.backend.domain.user.enums;
+
+public enum Gender {
+    MALE, FEMALE,
+}
