+++ apps/backend/src/main/java/com/sos/backend/domain/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.sos.backend.domain.user.repository;
+
+import com.sos.backend.domain.user.entity.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Long> {
+
+    Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
+}
